chromium-wayland: Define more boards as compatible
dragonboard-410c and m3ulcb explicitly set as compatible.
You would think that COMPATIBLE_MACHINE_armv8 = "(.*)" is more in line
with how the others are handled - even though that is not really
limiting any compatibility.  But this _armv8 suffix did not do the trick
it seems so I went for explicitly naming the boards for now.